+++ skbase/utils/dependencies/__init__.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python3 -u
+# -*- coding: utf-8 -*-
+# copyright: skbase developers, BSD-3-Clause License (see LICENSE file)
+"""Utility functionality used through `skbase`."""
+
+from skbase.utils.dependencies._dependencies import (
+    _check_python_version,
+    _check_soft_dependencies,
+)
+
+__all__ = ["_check_python_version", "_check_soft_dependencies"]
